refactor: import SearchQueryParams from useChannelSearch
